[MODEL] Chris gets his Longs
git-svn-id: file:///home/svn/omero/trunk@301 05709c45-44f0-0310-885b-81a1db45b4a6

+++ components/common/src/ome/api/NewModel.java
@@ -7,7 +7,7 @@ import ome.util.Filterable;
 
 public interface NewModel extends Filterable{
 	
-	public Integer getId();
+	public Long getId();
 	public Integer getVersion();
 	public Permissions getPermissions();
 	public Experimenter getOwner();
